fix navtree action reactivity
